chore : fix directory  structure
