+++ src/main/java/com/Techeer/Team_C/domain/user/jwt/JwtTokenProvider.java
@@ -1,10 +1,9 @@
-package com.Techeer.Team_C.User.domain.jwt;
+package com.Techeer.Team_C.domain.user.jwt;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
-import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
