+++ src/main/java/com/Techeer/Team_C/domain/user/repository/UserMemoryRepository.java
@@ -1,6 +1,6 @@
-package com.Techeer.Team_C.User.repository;
+package com.Techeer.Team_C.domain.user.repository;
 
-import com.Techeer.Team_C.User.domain.User;
+import com.Techeer.Team_C.domain.user.entity.User;
 import java.util.*;
 
 /*
